\name{Cancer}
\docType{data}
\alias{Cancer}
\title{Cancer Census Dataset}
\description{
  This example dataset contains a list of seven cancer predisposition gene sets.
}
\usage{Cancer}
\details{
The seven cancer predisposition gene sets are:

1) NRG (Rahman, N. Realizing the promise of cancer predisposition genes. Nature 2014, 505:302-308);

2) NBG (Tamborero, D. et al. Comprehensive identification of mutational cancer driver genes across 12 tumor types. Scientific reports 2013, 3:2650);

3) LDG (Kandoth, C. et al. Mutational landscape and significance across 12 major cancer types. Nature 2013, 502:333-339);

4) GGG (Lawrence, M. S. et al. Discovery and saturation analysis of cancer genes across 21 tumour types. Nature 2014, 505:495-501);

5) ELG (Garraway, L. A. & Lander, E. S. Lessons from the cancer genome. Cell 2013, 153:17-37);

6) CCG (Futreal, P. A. et al. A census of human cancer genes. Nature reviews. Cancer 2004, 4:177-183);

7) BVG (Vogelstein, B. et al. Cancer genome landscapes. Science 2013, 339:1546-1558).

}
\references{
Minghui Wang, Yongzhong Zhao, and Bin Zhang (2015). Efficient Test and Visualization of Multi-Set Intersections. \emph{Scientific Reports} 5: 16923.
}
\keyword{datasets}
\seealso{\code{\link{supertest}}}
