\name{cis.eqtls}
\docType{data}
\alias{cis.eqtls}
\title{cis-eQTLs}
\description{
  This example dataset contains a list of cis-eQTL genes.
}
\details{
A list is included in this dataset: \code{cis.eqtls}, which contains four sets of cis-eQTL genes published by Gibbs et al (PLOS Genetics 2010, 6:e1000952) as deposited in the eQTL Browser (http://www.ncbi.nlm.nih.gov/projects/gap/eqtl/index.cgi).
The four sets of cis-eQTL genes were detected in four different brain regions from Gibbs: brain cerebellum (CB), brain frontal cortex region (FC), brain temporal cortex region (TC), and brain pons region (PONS) respectively.
}
\keyword{datasets}
\seealso{\code{\link{supertest}}}
