\name{GWAS}
\docType{data}
\alias{GWAS}
\title{GAWS Catalog Dataset}
\description{
  This example dataset contains a list of gene sets associated with six types of clinical traits curated in the GWAS Catalog.
}
\usage{GWAS}
\details{
The six clinical traits are:

1) NEU (Bipolar disorder and schizophrenia, Schizophrenia, Major depressive disorder, Alzheimer's disease, Parkinson's disease, Cognitive performance, Bipolar disorder);

2) INF (Crohn's disease, Ulcerative colitis, Inflammatory bowel disease, Rheumatoid arthritis, Multiple sclerosis, Systemic lupus erythematosus);

3) CVD (Type 2 diabetes, Coronary heart disease, Blood pressure, total Cholesterol, HDL cholesterol, Triglycerides);

4) HT (height);

5) IgG (IgG glycosylation);

6) OB (obesity, obesity related traits).
}
\references{
Minghui Wang, Yongzhong Zhao, and Bin Zhang (2015). Efficient Test and Visualization of Multi-Set Intersections. \emph{Scientific Reports} 5: 16923.
}
\keyword{datasets}
\seealso{\code{\link{supertest}}}
