\name{SuperExactTest}
\alias{SuperExactTest}
\docType{package}
\title{SuperExactTest Package}
\description{
Efficient Test and Visualization of Multi-set Intersections
}
\details{
The main functions that most users may need from this package are \code{\link{supertest}} and \code{\link{MSET}}. For a brief introduction of using this package, please see \code{vignette("set_html")}.
}
\author{
Minghui Wang <minghui.wang@mssm.edu>, Bin Zhang <bin.zhang@mssm.edu>
}
\references{
Minghui Wang, Yongzhong Zhao, and Bin Zhang (2015). Efficient Test and Visualization of Multi-Set Intersections. \emph{Scientific Reports} 5: 16923.
}
\seealso{\code{\link{supertest}}, \code{\link{MSET}}}
\examples{
\dontrun{
#See a brieft instroduction of using this package
vignette("set_html")
}
}
