\name{GWAS}
\docType{data}
\alias{GWAS}
\title{GAWS Catalog Dataset}
\description{
  This example dataset contains a list of gene sets associated with six types of clinical traits curated in the GWAS Catalog.
}
\usage{data(GWAS)}
\details{
The six clinical traits are:
\itemize{
\item NEU (Bipolar disorder and schizophrenia, Schizophrenia, Major depressive disorder, Alzheimer's disease, Parkinson's disease, Cognitive performance, Bipolar disorder);
\item INF (Crohn's disease, Ulcerative colitis, Inflammatory bowel disease, Rheumatoid arthritis, Multiple sclerosis, Systemic lupus erythematosus);
\item CVD (Type 2 diabetes, Coronary heart disease, Blood pressure, total Cholesterol, HDL cholesterol, Triglycerides);
\item HT (height);
\item IgG (IgG glycosylation);
\item OB (obesity, obesity related traits).
}
}
\references{
Minghui Wang, Yongzhong Zhao, and Bin Zhang (2015). Efficient Test and Visualization of Multi-Set Intersections. \emph{Scientific Reports} 5: 16923.
}
\keyword{datasets}
\seealso{\code{\link{supertest}}}
